openrisc: add SMP and NR_CPUS Kconfig options

openrisc: select GENERIC_SMP_IDLE_THREAD


openrisc: spr_defs: add SPR_COREID and SPR_NUMCORES


openrisc: initial SMP support

Work-in-progress, contains stubs for the IPI stuff.
Apart from that secondary  bringup should be "working",
no guarantees though.
Boot loader spinning of second cpu is not supported yet,
it's assumed that Linux is booted straight after cpu reset.


openrisc: create a current_pgd for each cpu

The reason why this is created as an array, and not with
DEFINE_PER_CPU is that doing it this way, we'll save a load
in the tlb-miss handler (the load from __per_cpu_offset).


openrisc/smp: implement IPI

This imitates the arm64 implementation by having a
smp_cross_call callback that can be set by set_smp_cross_call
to initiate an IPI and a handle_IPI function that is expected
to be called from an IPI irqchip driver.


openrisc/smp: initialize current_pgd in __cpu_up


openrisc: add smp tlb flush functionality

The SMP implementation of flush_tlb_* works by sending out
a function-call IPI to all the non-local cpus by using the
generic on_each_cpu() function.

Currently, all flush_tlb_* functions will result in a
flush_tlb_all(), which has always been the behaviour in the
UP case.


openrisc/head: load r10 from per-cpu current_thread_info_set

In the SMP case, SPR_COREID is used to load the
correct cpu specific current_thread_info.
In the UP case, current_thread_info_set[0] is unconditionally
loaded.


openrisc: enable caches on secondary cpus

openrisc: improve description of OPENRISC_HAVE_SHADOW_GPRS


openrisc: add some missing spr registers


openrisc: make cpuinfo per-cpu specific

This creates a per cpu cpuinfo struct and fills it out
accordingly for each activated cpu.
show_cpuinfo is also updated to reflect new version information
in later versions of the spec.


openrisc: time: make timers work with smp

This is a humongous hack that takes an external free running
timer and synchronize the local timers to that.
I only commit it since the hack is a no-op on UP,
and it "works" on SMP.

The proper thing to do here is to completely use an external
timer with seperate capture interrupt channels.


openrisc: delay: use get_cycles() instead of reading the raw timer


openrisc: smp: init clockevent on secondary cpus


openrisc: Configure clockevents broadcast

Build failed without

openrisc: smp: move startup to init section

Add __init annotation, the compile had a warning about this.

openrisc: time: Fix error with invalid c

Cant declare and assign in same line with kernel c.

openrisc: fix build warnings

unused var in smp.c
bad type of the vmalloc_start/end causing:

  mm/percpu.c: In function 'pcpu_embed_first_chunk':
  ./include/linux/kern_levels.h:4:18: warning: format '%lx' expects
  argument of type 'long unsigned int', but argument 3 has type 'unsigned int'

TODO get review + merge

+++ arch/openrisc/include/asm/pgtable.h
@@ -94,8 +94,8 @@ extern void paging_init(void);
  * 64 MB of vmalloc area is comparable to what's available on other arches.
  */
 
-#define VMALLOC_START	(PAGE_OFFSET-0x04000000)
-#define VMALLOC_END	(PAGE_OFFSET)
+#define VMALLOC_START	(unsigned long)(PAGE_OFFSET-0x04000000)
+#define VMALLOC_END	(unsigned long)(PAGE_OFFSET)
 #define VMALLOC_VMADDR(x) ((unsigned long)(x))
 
 /* Define some higher level generic page attributes.
